cli: move `wallet candidate getstate` into `query voter`
This command has nothing to do with wallets.
